dr version 0.2: edit expressions!
I'd like to also merge find-tagged / open-tagged into this, and I don't
think it'll be all that difficult.

I'm questioning the wisdom of switching from if/elif to case; there may
be more regex matches in the future of edit expressions, and an
if-elif-else case-esac-fi expression is a hair awkward. But then again
monster chains of [[ && ]] conditions aren't exactly pretty either.

Removed 'random' and 'last' commands, as their functions are completely
subsumed by edit expressions ('dr edit -s *' and 'dr edit last',
respectively).