improve search documentation; rename 'setdiff'; add NOT
I wasn't thinking very hard when I named "setdiff" -- while that kind of
gets the idea of what's going on, set difference is NOT XOR (heh heh),
so that was confusing. Now called 'xwinnow'.

You can now invert any match with -v; I honestly doubt this is usually
going to be terribly useful, but it seems proper to support all boolean
operators in some way.