C implementation of word count
This runs in a tiny fraction of the time of the Python implementation --
about 3% by my count. Over 1200 dreams, the Python one took several
seconds, whereas this feels almost instantaneous.

Of course, now I have to figure out a build system of some kind since I
have C code, and ideally there would be a fallback to the Python
implementation if it hasn't been compiled.

Also, it should be noted this doesn't necessarily *quite* behave right
with multi-byte characters; they might result in slight miscounts if
isalphanum() doesn't return the right answer for part of a multi-byte
character. This doesn't create much of an issue for the occasional
umlauts or accents that I use (the counts were literally only off by a
few words over the entire dreamdir of nearly 300K words), but if you're
writing your dreams in e.g., Japanese or Hebrew, this implementation
might be essentially worthless.