more word counting
You can now choose between two output formats for the word count
(machine and human-readable), as well as which of the word count types
you want to see.

The Python word-count routines have been updated to have the same
options and output possibilities (along with a Python dictionary return
for use in other scripts). Of course, they're still way slower.

A Makefile has been added to build the binary. A small stub has been
added to 'dr' to invoke the word count for a search-expr with the '-p'
option; the ability to pass additional options along to the word count
program would be nice in the future. If the binary is not available,
'dr' will fall back on the Python implementation.