github issue #3: find regex fails if no leading 0s in dream #
Not likely to come up for quite some time (so I'm glad I caught it).
This gets away with fewer sed commands in the pipeline than before, too.

Also add note about how to get more help to the error you get if
'dreamfind' finds a file that doesn't exist.