fix RegeneratePlots.sh failing with odd error
The file paths were wacky because the scripts/ directory in my dreamdir
is symlinked into my development tree. Thus '../graphs' didn't lead back
out to the graphs folder in my actual dreamdir.

To avoid this being a problem for me or anyone else, we now work only
from the root of the dreamdir in RegeneratePlots.sh at all times.