bugfix: limit not working
I missed the 'tr' command that was used in the two very similar
pipelines immediately before it. Not sure how I didn't notice this
before -- it must have been working in a previous version.

A test has been added for this (though the tests are still not
anywhere close to complete, I figure I might as well throw in a test of
something that was broken now and add other similar ones later).