improvements to titleadd script
You can now choose what header you want to add, and the choice has
changed to "display headers automatically" or "display whole dream
automatically," as not displaying the headers at all is useful only in
very rare circumstances.

Also, the output of 'ed' no longer appears on the screen when a file is
changed.