improve header-replace runtime spectacularly
We now filter out any dreams that don't match the search pattern (using
'dreamfind tagged') so that we only have to run the (rather slow)
replace loop over dreams that are actually going to match the pattern.

Also change & improve the instructions / output presented to the user
during the process.