permit blank search-exprs; use search-expr for header-replace
These are identical to 'all', and are particularly nice with the recent
addition of search-expr syntax to several functions that one typically
applies to everything. Since search-exprs always come last in the
argument list, you can just leave them off if you want to match
everything.

header-replace previously used a list of filenames, which was kind of
silly.