shellcheck: add quotes to some things
...and then have to disable the quote check for the whole script for the
future because it flags a lot of things that would break the script if
quoted (like globs in variables).

I know I could use (and the docs suggest using) an array instead, but
expanding a glob into an array with hundreds or thousands of elements
only to then copy out the entire array into the command-line arguments
of another command is unnecessarily verbose and useless overhead.