fix issue #17: "lucid dreams today" stat wrong
This was inadvertently introduced in commit aab92b96, when I tried using
'grep -c' as suggested by shellcheck, but it isn't actually what's
needed here. I evidently forgot to put it back after I decided it wasn't
useful. Amusingly, the line above is a comment reading "not an
appropriate usage of 'grep -c'".