don't find dream numbers when tag value is numeric
This resolves #16.

The bug occurred because the tag search procedure runs its search on a
line something like this (for a search on the Tags header):

    00911.dre:@@@@@theft@@@@@bicycles@@@@@false awakening@@@@@computer\
    @@@@@marriage@@@@@library@@@@@Fate@@@@@money@@@@@

A search 'tagged Tags 911' will match the '00911.dre' part and mark this
dream as a match.

To correct this, the part of the line before the colon is now matched
explicitly so that the user's pattern can't match it.