work around broken terminal on 'dr act edit'
Since stdin is being redirected within the command, vim gives a warning
and then your terminal is broken when you exit due to (presumably)
curses or whatever terminal library not knowing how to restore the
settings properly. I don't understand quite enough about redirection to
explain why exactly you can still use the editor even though it acts up
a bit, but it does make a vague kind of sense to me. This is also a
problem with, for instance, xargs. Fortunately, it's pretty easy to fix
-- you just have to explicitly redirect the stdin of the editor from
/dev/tty.