don't act up (heh) when no data reaches 'dr act'
There was an error already raised at the point of reaching the first
'winnow' in a pipeline that didn't start with 'filename-display', but
the other parts of the pipeline were still kicked off by the OS, and 'dr
act' ended up reading no arguments and simply performing its action on
the entire dreamdir (since its input was blank and a blank
search-expression indicates everything). 'dr act' now requires some
input or it will die itself.

Closes #18.