overhaul word-count
We now support only 'drwc', but you don't have to install it if you
don't want to use the functionality I suppose. This simplifies matters
considerably; now we can just pass through arguments to there. I added
an issue for the incorrect handling of multibyte characters and we'll
get to fixing that at some point in the future.

'drwc' also now has to be on the system path. This removes our last
dependency on the 'scripts' directory.

Closes #14 since we don't have two different versions anymore.