remove GNU grep-specific switch
'-m 1' stops searching after the first match. This is somewhat a
GNU-specific extension. In the MacOS BSD grep (2.5.1-FreeBSD), '-m' is
listed as supported in the manual with the explanation "Stop reading the
file after _num_ matches,", but behaves differently than the GNU version
and definitely violates the Principle of Least Astonishment with
multiple files: it treats the count as applying to *all* files as if
they were concatenated. This was resulting in only the first match ever
appearing.

At any rate, we shouldn't be using grep switches that aren't defined by
the POSIX standard, as we aren't listing any specific grep as a
dependency..

Since this was a performance optimization rather than necessary for
correctness, I have simply removed it. If this proves to be an actual
performance issue later on, we can experiment with using awk or sed to
get this behavior instead.