fix tac/shuf wrappers
The 'hash' bash builtin seems to treat a function whose definition is
in progress as a match for the tool that's being tested, which is rather
odd behavior and leads to the incorrect conclusion that 'tac' is
installed on my Mac. We will simply rename shuf and tac to 'shufw' and
'tacw' (for 'tac wrapper') and avoid the conflict.