don't create backup files with '-e' extension
BSD sed is incompatible with GNU sed in a really irritating way: -i is
required to take an empty argument on BSD sed to skip backup file
creation but not allowed to take an empty argument on GNU sed. Using a
wrapper function obtained from SO to fix this issue.